+++ test/testdata/gofumpt.go
@@ -1,0 +1,11 @@
+//args: -Egofumpt
+package testdata
+
+import "fmt"
+
+func GofumptNewLine() {
+
+	fmt.Println("foo")
+}
+
+// ERROR "File is not `gofumpt`-ed"